finalize cmd/create helper app ; no more src/ dir
duplicate all "common" code from src/ to the apps in cmd/
where they are used. it's not that much code and there are minor
differences in implementation between each cmd/ app so it's better this way
